Move legacy to root
