+++ react/legacy/Footer.js
