+++ react/legacy/FooterBadgeList.js
