+++ react/legacy/footerList.js
